Upgrade agent to 2.0.4

+++ bootstrap.sh
@@ -42,13 +42,13 @@ BUILDKITE_PROMPT="\033[90m$\033[0m"
 # Shows the command being run, and runs it
 function buildkite-prompt-and-run {
   echo -e "$BUILDKITE_PROMPT $1"
-  eval $1
+  eval "$1"
 }
 
 # Shows the command about to be run, and exits if it fails
 function buildkite-run {
   echo -e "$BUILDKITE_PROMPT $1"
-  eval $1
+  eval "$1"
   EVAL_EXIT_STATUS=$?
 
   if [[ $EVAL_EXIT_STATUS -ne 0 ]]; then
@@ -65,7 +65,7 @@ function buildkite-debug {
 # Runs the command, but only output what it's doing if we're in DEBUG mode
 function buildkite-run-debug {
   buildkite-debug "$BUILDKITE_PROMPT $1"
-  eval $1
+  eval "$1"
 }
 
 # Show an error and exit
@@ -73,6 +73,12 @@ function buildkite-error {
   echo -e "~~~ :rotating_light: \033[31mBuildkite Error\033[0m"
   echo "$1"
   exit 1
+}
+
+# Show a warning
+function buildkite-warning {
+  echo -e "\033[33m⚠️ Buildkite Warning: $1\033[0m"
+  echo "^^^ +++"
 }
 
 # Run a hook script
@@ -97,7 +103,7 @@ function buildkite-hook {
     buildkite-flags-reset
 
     # Return back to the working dir
-    cd $HOOK_PREVIOUS_WORKING_DIR
+    cd "$HOOK_PREVIOUS_WORKING_DIR"
 
     # Exit from the bootstrap.sh script if the hook exits with a non-0 exit
     # status
@@ -131,7 +137,7 @@ function buildkite-local-hook {
 export PATH="$BUILDKITE_BIN_PATH:$PATH"
 
 # Come up with the place that the repository will be checked out to
-SANITIZED_AGENT_NAME=$(echo $BUILDKITE_AGENT_NAME | tr -d '"')
+SANITIZED_AGENT_NAME=$(echo "$BUILDKITE_AGENT_NAME" | tr -d '"')
 PROJECT_FOLDER_NAME="$SANITIZED_AGENT_NAME/$BUILDKITE_PROJECT_SLUG"
 export BUILDKITE_BUILD_CHECKOUT_PATH="$BUILDKITE_BUILD_PATH/$PROJECT_FOLDER_NAME"
 
@@ -185,16 +191,16 @@ else
     # Only bother running the keyscan if the SSH host has been provided by
     # Buildkite. It won't be present if the host isn't using the SSH protocol
     if [[ ! -z "${BUILDKITE_REPO_SSH_HOST:-}" ]]; then
-      : ${BUILDKITE_SSH_DIRECTORY:="$HOME/.ssh"}
-      : ${BUILDKITE_SSH_KNOWN_HOST_PATH:="$BUILDKITE_SSH_DIRECTORY/known_hosts"}
+      : "${BUILDKITE_SSH_DIRECTORY:="$HOME/.ssh"}"
+      : "${BUILDKITE_SSH_KNOWN_HOST_PATH:="$BUILDKITE_SSH_DIRECTORY/known_hosts"}"
 
       # Ensure the known_hosts file exists
-      mkdir -p $BUILDKITE_SSH_DIRECTORY
-      touch $BUILDKITE_SSH_KNOWN_HOST_PATH
+      mkdir -p "$BUILDKITE_SSH_DIRECTORY"
+      touch "$BUILDKITE_SSH_KNOWN_HOST_PATH"
 
       # Only add the output from ssh-keyscan if it doesn't already exist in the
       # known_hosts file
-      if ! ssh-keygen -H -F "$BUILDKITE_REPO_SSH_HOST" | grep --quiet "$BUILDKITE_REPO_SSH_HOST"; then
+      if ! ssh-keygen -H -F "$BUILDKITE_REPO_SSH_HOST" | grep -q "$BUILDKITE_REPO_SSH_HOST"; then
         buildkite-run "ssh-keyscan \"$BUILDKITE_REPO_SSH_HOST\" >> \"$BUILDKITE_SSH_KNOWN_HOST_PATH\""
       fi
     fi
@@ -209,7 +215,10 @@ else
   fi
 
   buildkite-run "git clean -fdq"
-  buildkite-run "git submodule foreach --recursive git clean -fdq"
+
+  if [[ -z "${BUILDKITE_DISABLE_GIT_SUBMODULES:-}" ]]; then
+    buildkite-run "git submodule foreach --recursive git clean -fdq"
+  fi
 
   buildkite-run "git fetch -q"
 
@@ -219,22 +228,35 @@ else
     buildkite-run "git fetch origin \"+refs/pull/$BUILDKITE_PULL_REQUEST/head:\""
   elif [[ "$BUILDKITE_TAG" == "" ]]; then
     # Default empty branch names
-    : ${BUILDKITE_BRANCH:=master}
+    : "${BUILDKITE_BRANCH:=master}"
 
     buildkite-run "git reset --hard origin/$BUILDKITE_BRANCH"
   fi
 
   buildkite-run "git checkout -qf \"$BUILDKITE_COMMIT\""
 
-  # `submodule sync` will ensure the .git/config matches the .gitmodules file
-  buildkite-run "git submodule sync --recursive"
-  buildkite-run "git submodule update --init --recursive"
-  buildkite-run "git submodule foreach --recursive git reset --hard"
+  if [[ -z "${BUILDKITE_DISABLE_GIT_SUBMODULES:-}" ]]; then
+    # `submodule sync` will ensure the .git/config matches the .gitmodules file.
+    # The command is only available in git version 1.8.1, so if the call fails,
+    # continue the bootstrap script, and show an informative error.
+    buildkite-prompt-and-run "git submodule sync --recursive"
+    if [[ $? -ne 0 ]]; then
+      buildkite-warning "Failed to recursively sync git submodules. This is most likely because you have an older version of git installed ($(git --version)) and you need version 1.8.1 and above. If you're using submodules, it's highly recommended you upgrade if you can."
+    fi
+
+    buildkite-run "git submodule update --init --recursive"
+    buildkite-run "git submodule foreach --recursive git reset --hard"
+  fi
 
   # Grab author and commit information and send it back to Buildkite
   buildkite-debug "~~~ Saving Git information"
-  buildkite-run-debug "buildkite-agent meta-data set \"buildkite:git:commit\" \"\`git show \"$BUILDKITE_COMMIT\" -s --format=fuller --no-color\`\""
-  buildkite-run-debug "buildkite-agent meta-data set \"buildkite:git:branch\" \"\`git branch --contains \"$BUILDKITE_COMMIT\" --no-color\`\""
+
+  # Check to see if the meta data exists before setting it
+  buildkite-run-debug "buildkite-agent meta-data exists \"buildkite:git:commit\""
+  if [[ $? -ne 0 ]]; then
+    buildkite-run-debug "buildkite-agent meta-data set \"buildkite:git:commit\" \"\`git show \"$BUILDKITE_COMMIT\" -s --format=fuller --no-color\`\""
+    buildkite-run-debug "buildkite-agent meta-data set \"buildkite:git:branch\" \"\`git branch --contains \"$BUILDKITE_COMMIT\" --no-color\`\""
+  fi
 fi
 
 # Store the current value of BUILDKITE_BUILD_CHECKOUT_PATH, so we can detect if
@@ -277,9 +299,13 @@ BUILDKITE_SCRIPT_PATH="buildkite-script-$BUILDKITE_JOB_ID"
 # Generate a different script depending on whether or not it's a script to
 # execute
 if [[ -f "$BUILDKITE_COMMAND" ]]; then
-  echo -e '#!/bin/bash'"\nchmod +x \"$BUILDKITE_COMMAND\"\n./\"$BUILDKITE_COMMAND\"" > $BUILDKITE_SCRIPT_PATH
+  # Make sure the script they're trying to execute has chmod +x. We can't do
+  # this inside the script we generate because it fails within Docker:
+  # https://github.com/docker/docker/issues/9547
+  buildkite-run-debug "chmod +x \"$BUILDKITE_COMMAND\""
+  echo -e '#!/bin/bash'"\n./\"$BUILDKITE_COMMAND\"" > "$BUILDKITE_SCRIPT_PATH"
 else
-  echo -e '#!/bin/bash'"\n$BUILDKITE_COMMAND" > $BUILDKITE_SCRIPT_PATH
+  echo -e '#!/bin/bash'"\n$BUILDKITE_COMMAND" > "$BUILDKITE_SCRIPT_PATH"
 fi
 
 if [[ "$BUILDKITE_AGENT_DEBUG" == "true" ]]; then
@@ -328,8 +354,8 @@ if [[ -e "$BUILDKITE_HOOKS_PATH/command" ]]; then
 else
   ## Docker
   if [[ ! -z "${BUILDKITE_DOCKER:-}" ]] && [[ "$BUILDKITE_DOCKER" != "" ]]; then
-    DOCKER_CONTAINER="buildkite_"$BUILDKITE_JOB_ID"_container"
-    DOCKER_IMAGE="buildkite_"$BUILDKITE_JOB_ID"_image"
+    DOCKER_CONTAINER="buildkite_${BUILDKITE_JOB_ID}_container"
+    DOCKER_IMAGE="buildkite_${BUILDKITE_JOB_ID}_image"
 
     function docker-cleanup {
       echo "~~~ Cleaning up Docker containers"
@@ -364,7 +390,7 @@ else
       buildkite-run "$COMPOSE_COMMAND rm --force -v || true"
 
       # The adhoc run container isn't cleaned up by compose, so we have to do it ourselves
-      buildkite-run "docker rm -f -v "$COMPOSE_CONTAINER_NAME"_run_1 || true"
+      buildkite-run "docker rm -f -v ${COMPOSE_CONTAINER_NAME}_run_1 || true"
     }
 
     trap compose-cleanup EXIT
@@ -417,10 +443,22 @@ if [[ "$BUILDKITE_ARTIFACT_PATHS" != "" ]]; then
 
   echo "~~~ Uploading artifacts"
   if [[ ! -z "${BUILDKITE_ARTIFACT_UPLOAD_DESTINATION:-}" ]] && [[ "$BUILDKITE_ARTIFACT_UPLOAD_DESTINATION" != "" ]]; then
-    buildkite-run "buildkite-agent artifact upload \"$BUILDKITE_ARTIFACT_PATHS\" \"$BUILDKITE_ARTIFACT_UPLOAD_DESTINATION\""
+    buildkite-prompt-and-run "buildkite-agent artifact upload \"$BUILDKITE_ARTIFACT_PATHS\" \"$BUILDKITE_ARTIFACT_UPLOAD_DESTINATION\""
   else
-    buildkite-run "buildkite-agent artifact upload \"$BUILDKITE_ARTIFACT_PATHS\""
+    buildkite-prompt-and-run "buildkite-agent artifact upload \"$BUILDKITE_ARTIFACT_PATHS\""
   fi
+
+  # If the artifact upload fails, open the current group and exit with an error
+  if [[ $? -ne 0 ]]; then
+    echo "^^^ +++"
+    exit 1
+  fi
+  
+  # Run the per-checkout `post-artifact` hook
+  buildkite-local-hook "post-artifact"
+
+  # Run the global `post-artifact` hook
+  buildkite-global-hook "post-artifact"
 fi
 
 # Be sure to exit this script with the same exit status that the users build
